requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-chroma
pdfplumber
duckduckgo-search
python-dotenv
ddgs